+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- new parameter called `allow-archived-tags` to allow selection of archived tags.
+
 ## [v0.40.0] - 2022-08-09
 
 ### Added